% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{calculate_coltheartN}
\alias{calculate_coltheartN}
\title{Function calculates the orthographic neighbours of
a list of words using (max_cores - 1) cores.}
\usage{
calculate_coltheartN(targets, all_words)
}
\arguments{
\item{targets}{target strings}

\item{all_words}{all words}
}
\value{
list of the number of neighbours
}
\description{
Function calculates the orthographic neighbours of
a list of words using (max_cores - 1) cores.
}
