% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{coltheartN}
\alias{coltheartN}
\title{Function calculates the number of orthographic neighbours of a single string}
\usage{
coltheartN(target_str, all_words, showN = FALSE)
}
\arguments{
\item{target_str}{target character string}

\item{all_words}{list of words}

\item{showN}{show neighours}
}
\value{
number of neighbours
}
\description{
Function calculates the number of orthographic neighbours of a single string
}
\examples{
\dontrun{
coltheartN("book", subtlex_uk$Spelling)
}
}
