% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{is_neighbour}
\alias{is_neighbour}
\title{strsim}
\usage{
is_neighbour(w1, w2)
}
\arguments{
\item{w1}{character string}

\item{w2}{character string}
}
\value{
boolean
}
\description{
Function to check that two strings are orthographic neighbours
}
\examples{
\dontrun{
# Orthographic Neighbour
is_neighbour("book", "look")
# Not a neighbour
is_neighbour("book", "boll")
# Not a neighbour
is_neighbour("book", "boo")
# Not a neighbour
is_neighbour("book", "boök")
}

}
