% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{nd}
\alias{nd}
\title{Neighbourhood Density}
\usage{
nd(target_str, dt, show = FALSE)
}
\arguments{
\item{target_str}{target character string}

\item{dt}{data table with words}

\item{show}{show neighours}
}
\value{
number of neighbours
}
\description{
`nd()` calculates the neighbourhood density of a string. It is based
on the definition of an orthographic neighbour by Coltheart et al. (1977).
}
\examples{
\dontrun{
nd("book", data.table(word = all_words))
}
}
