% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{nd_string}
\alias{nd_string}
\title{Function calculates neighbourhood density of a single string. It is based
on the definition of a neighbour by Coltheart et al. (1977).}
\usage{
nd_string(target_str, dt, show = FALSE)
}
\arguments{
\item{target_str}{target character string}

\item{dt}{data table with words}

\item{show}{show neighours}
}
\value{
number of neighbours
}
\description{
Function calculates neighbourhood density of a single string. It is based
on the definition of a neighbour by Coltheart et al. (1977).
}
\examples{
\dontrun{
nd_string("book", data.table(word = all_words))
}
}
