% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{neighborhood_density}
\alias{neighborhood_density}
\title{Function calculates the neighbourhood density of
a list of words.}
\usage{
neighborhood_density(targets_spellings, all_words, show = FALSE)
}
\arguments{
\item{targets_spellings}{target strings}

\item{all_words}{all words}

\item{show}{show neighbours}
}
\value{
list of the number of neighbours
}
\description{
Function calculates the neighbourhood density of
a list of words.
}
\examples{
\dontrun{
neighborhood_density("book", subtlex_uk$Spelling)
neighborhood_density(subtlex_uk$Spelling, subtlex_uk$Spelling)
}
}
