% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{nf}
\alias{nf}
\title{Neighbourhood Frequency}
\usage{
nf(target, dt_corpus, show = FALSE)
}
\arguments{
\item{target}{target string}

\item{dt_corpus}{data table with words and frequencies (word, frequency)}

\item{show}{show higher frequency neighours}
}
\value{
number of higher frequency neighbours
}
\description{
`nf()` calculates the neighbourhood frequency of a single string. Frequency
of the target string is looked up in the data table `dt_corpus`.
}
\examples{
\dontrun{
library(data.table)
dt <- data.table( word = subtlex_uk$Spelling, frequency = subtlex_uk$`LogFreq(Zipf)`)
nf("book", dt, show=T)
}
}
