% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{nf_string}
\alias{nf_string}
\title{Function calculates the neighbourhood frequency of a single string. Frequency
of that target string is looked up in the data table.}
\usage{
nf_string(target, dt_corpus, show = FALSE)
}
\arguments{
\item{target}{target string}

\item{dt_corpus}{data table with words and frequencies (word, frequency)}

\item{show}{show higher frequency neighours}
}
\value{
number of higher frequency neighbours
}
\description{
Function calculates the neighbourhood frequency of a single string. Frequency
of that target string is looked up in the data table.
}
\examples{
\dontrun{
library(data.table)
dt <- data.table( word = subtlex_uk$Spelling, frequency = subtlex_uk$`LogFreq(Zipf)`)
nf_string("book", dt, show=T)
}
}
