% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{old}
\alias{old}
\title{Orthographic Levenshtein Distance}
\usage{
old(target_word, words, old_n = 20, show = FALSE)
}
\arguments{
\item{target_word}{target word}

\item{words}{vector of words}

\item{old_n}{number of words to calculate old20, default is 20}

\item{show}{show the Levenshtein distance of each of the n_old (20) words that have
the lowest Levenshtein distance with the target word.}
}
\value{
average Levenshtein distance
}
\description{
`old()` calculates the average Orthographic Levenshtein Distance (OLD)
of a single string. When n_old is 20 (default value), the function
calculates OLD20 as defined by
\href{http://link.springer.com/article/10.3758/PBR.15.5.971}{Yarkoni et al. (2008)}.
}
\examples{
\dontrun{
old("book", subtlex_uk$Spelling)
}
}
