% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{calculate_coltheartN}
\alias{calculate_coltheartN}
\title{Function uses multiple cores to calculate the orthographic neighbours of
a list of words}
\usage{
calculate_coltheartN(targets, all_words)
}
\arguments{
\item{targets}{target strings}

\item{all_words}{all words}
}
\value{
list of the number of neighbours
}
\description{
Function uses multiple cores to calculate the orthographic neighbours of
a list of words
}
