% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{calculate_old20}
\alias{calculate_old20}
\title{Function calculates OLD20 of a list of strings using multiple cores}
\usage{
calculate_old20(target_words, all_words, old_n = 20)
}
\arguments{
\item{all_words}{list of all words}

\item{old_n}{max number of most similar words to calculate OLD20}

\item{target_word}{target string}
}
\value{
list of words with OLD20 values
}
\description{
Function calculates OLD20 of a list of strings using multiple cores
}
