% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{is_neighbour}
\alias{is_neighbour}
\title{Are two strings neighbours?}
\usage{
is_neighbour(w1, w2)
}
\arguments{
\item{w1}{character string}

\item{w2}{character string}
}
\value{
boolean
}
\description{
`is_neighbour()` checks whether two strings are orthographic neighbours
according to Coltheart et al.'s (1977) definition of a neighbour (Coltheart's N).
}
\examples{
{
# Orthographic Neighbour
is_neighbour("book", "look")
# Not a neighbour
is_neighbour("book", "boll")
# Not a neighbour
is_neighbour("book", "boo")
# Orthographic Neighbour
is_neighbour("book", "boök")
}

}
