% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{neighborhood_density}
\alias{neighborhood_density}
\title{Neighbourhood Density of a list of words}
\usage{
neighborhood_density(
  targets_spellings,
  all_words,
  show = FALSE,
  pb = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{targets_spellings}{target strings}

\item{all_words}{all words}

\item{show}{show neighbours}

\item{pb}{show progress bar}

\item{parallel}{parallel processing}
}
\value{
list of the number of neighbours
}
\description{
neighborhood_density()` calculates the neighbourhood density of a list of words.
}
\examples{
\dontrun{
nd("book", subtlex_uk$Spelling)
nd(subtlex_uk$Spelling, subtlex_uk$Spelling)
}
}
