% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsim.R
\name{old20}
\alias{old20}
\title{Function that calculate the OLD20 of a single string}
\usage{
old20(target_word, words, old_n = 20, showN = FALSE)
}
\arguments{
\item{target_word}{target string}

\item{words}{list of words}

\item{old_n}{number of words to calculate old20, default is 20}

\item{showN}{show the orthographic similar words}
}
\value{
mean levenshtein distance
}
\description{
OLD20 defintion can be found in Yarkoni et al. (2008), see
http://link.springer.com/article/10.3758/PBR.15.5.971
}
